MNT: rename the project and package to shed-streaming.
