+++ shed_streaming/writers.py
